requirements.txt
urllib
img2pdf
bs4
Pillow
python-resize-image